# app/blueprints/service_tickets/schemas.py
from app.extensions import ma
from app.models import ServiceTickets, db


class ServiceTicketSchema(ma.SQLAlchemyAutoSchema):
    class Meta:
        model = ServiceTickets 
        include_fk=True

service_ticket_schema = ServiceTicketSchema() 
service_tickets_schema = ServiceTicketSchema(many=True) 